requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk==3.5.0
requests
selenium
webdriver-manager
chromedriver-binary==122.0.6259.0.0
